Refactor: 설정 페이지 컴포넌트 분리 (#26)
* Modify: 유저 이미지 파트 폴더명 수정

* Modify: 유저 이미지 파트 폴더명 수정

* Modify: 유저 이미지 css 순서 변경

* Modify: 유저 소개 폴더명 수정

* Refactor: 회원탈퇴 컴포넌트 분리

* Refactor: 회원탈퇴 컴포넌트 분리

* Refactor: 이메일 수신 설정 컴포넌트 분리

* Refactor: 이메일과 소셜 정보 컴포넌트 분리

* Refactor: 벨로그 제목 컴포넌트 분리

* Refactor: 벨로그 제목과 소셜 정보 수정 함수 이름 변경

* Modify: 소셜 정보 수정 시 스타일 수정

* Modify: 소셜 정보 수정 시 input margin-top 수정

* Refactor: 유저 프로필 이미지 컴포넌트 분리

* Refactor: 유저 프로필 이미지 컴포넌트 분리

* Refactor: 설정 페이지 반응형 변수로 관리
