+++ src/components/Setting/UserProfile/Index.jsx
@@ -1,12 +1,13 @@
+import UserProfileImage from './UserProfileImage';
 import Button from '../../Button';
-import profile from '../../../assets/profile.png';
 import styled from 'styled-components';
 import { buttonText, primary1, primary2, slightLayer } from '../../../styles/color';
+import { settingProfileMaxWidth768px, settingProfileButtonMaxWidth768px } from '../../../styles/media';
 
 const UserProfile = () => {
   return (
     <UserProfileContainer>
-      <img alt='profile' src={profile} />
+      <UserProfileImage />
       <Button text='이미지 업로드' className='upload' />
       <Button text='이미지 제거' className='remove' />
     </UserProfileContainer>
@@ -18,25 +19,10 @@ const UserProfileContainer = styled.div`
   flex-direction: column;
   padding-right: 1.5rem;
 
-  img {
-    object-fit: cover;
-    width: 8rem;
-    height: 8rem;
-    border-radius: 50%;
-    display: block;
-    margin-bottom: 1.25rem;
-
-    @media screen and (max-width: 768px) {
-      width: 6rem;
-      height: 6rem;
-      margin-bottom: 1rem;
-    }
-  }
+  ${settingProfileMaxWidth768px};
 
   button {
-    @media screen and (max-width: 768px) {
-      width: 10rem;
-    }
+    ${settingProfileButtonMaxWidth768px};
   }
 
   .upload {
@@ -61,13 +47,6 @@ const UserProfileContainer = styled.div`
   button + button {
     margin-top: 0.5rem;
     margin-left: 0px;
-  }
-
-  @media (max-width: 768px) {
-    -webkit-box-align: center;
-    align-items: center;
-    padding-bottom: 1.5rem;
-    padding-right: 0px;
   }
 `;
