+++ src/pages/Setting/Index.jsx
@@ -1,17 +1,17 @@
 import UserProfile from '../../components/Setting/UserProfile';
-import UserInfo from '../../components/Setting/UserInfo';
-import UserContents from '../../components/Setting/UserContents';
+import UserIntroduction from '../../components/Setting/UserIntroduction';
+import UserInformation from '../../components/Setting/UserInformation';
 import styled from 'styled-components';
-import { settingMaxWidth1024px, settingMaxWidth768px } from '../../styles/media';
+import { settingMaxWidth1024px, settingMaxWidth768px, settingUserMaxWidth768px } from '../../styles/media';
 
 const Setting = () => {
   return (
     <SettingPage>
-      <section className='setting-top'>
+      <section className='setting-user'>
         <UserProfile />
-        <UserInfo />
+        <UserIntroduction />
       </section>
-      <UserContents />
+      <UserInformation />
     </SettingPage>
   );
 };
@@ -26,14 +26,11 @@ const SettingPage = styled.div`
   ${settingMaxWidth1024px};
   ${settingMaxWidth768px};
 
-  .setting-top {
+  .setting-user {
     display: flex;
     height: 13.75rem;
 
-    @media (max-width: 768px) {
-      height: auto;
-      flex-direction: column;
-    }
+    ${settingUserMaxWidth768px};
   }
 `;
