+++ src/styles/media.js
@@ -28,8 +28,47 @@ export const settingMaxWidth1024px = `@media screen and (max-width: 1024px) {
   padding-left: 1rem;
   padding-right: 1rem;
 }`;
-
 export const settingMaxWidth768px = `@media screen and (max-width: 768px) {
   width: 100%;
   margin-top: 1.5rem;
+}`;
+export const settingUserMaxWidth768px = `@media (max-width: 768px) {
+  height: auto;
+  flex-direction: column;
+}`;
+
+export const settingProfileMaxWidth768px = `@media (max-width: 768px) {
+  -webkit-box-align: center;
+  align-items: center;
+  padding-bottom: 1.5rem;
+  padding-right: 0px;
+}`;
+export const settingProfileButtonMaxWidth768px = `@media screen and (max-width: 768px) {
+  width: 10rem;
+}`;
+export const UserProfileImageMaxWidth768px = `@media screen and (max-width: 768px) {
+  width: 6rem;
+  height: 6rem;
+  margin-bottom: 1rem;
+}`;
+import { border4 } from './color';
+export const UserIntroMaxWidth768px = `@media screen and (max-width: 768px) {
+  padding-top: 1.5rem;
+  padding-bottom: 1.5rem;
+  border-top: 1px solid ${border4};
+  border-bottom: 1px solid ${border4};
+  border-left: none;
+  padding-left: 0px;
+}`;
+export const UserIntroTitleMaxWidth768px = `@media screen and (max-width: 768px) {
+  font-size: 1.25rem;
+}`;
+export const UserContentsMaxWidth768px = `@media screen and (max-width: 768px) {
+  margin-top: 0rem;
+}`;
+export const UserContentsBoxMaxWidth768px = `@media screen and (max-width: 768px) {
+  flex-direction: column;
+}`;
+export const UserContentsTitleMaxWidth768px = `@media screen and (max-width: 768px) {
+  margin-bottom: 0.5rem;
 }`;
